move a few tests
